+++ src/components/ui/forms/ABISelector.tsx
@@ -1,21 +1,8 @@
 import { Select, Text } from '@chakra-ui/react';
-import axios from 'axios';
-import detectProxyTarget from 'evm-proxy-detection';
-import { useEffect, useMemo, useState } from 'react';
+import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import { isAddress } from 'viem';
-import { logError } from '../../../helpers/errorLogging';
-import { useNetworkEnsAddress } from '../../../hooks/useNetworkEnsAddress';
-import useNetworkPublicClient from '../../../hooks/useNetworkPublicClient';
-import { useNetworkConfigStore } from '../../../providers/NetworkConfig/useNetworkConfigStore';
+import { ABIElement, useABI } from '../../../hooks/utils/useABI';
 import { LabelComponent } from './InputComponent';
-
-export type ABIElement = {
-  type: 'function' | 'constructor' | 'fallback' | 'event';
-  name: string;
-  stateMutability: 'view' | 'nonpayable' | 'pure';
-  inputs: { type: string; name: string; internalType: string }[];
-};
 
 interface IABISelector {
   /*
@@ -26,45 +13,8 @@ interface IABISelector {
 }
 
 export default function ABISelector({ target, onChange }: IABISelector) {
-  const [abi, setABI] = useState<ABIElement[]>([]);
-  const { etherscanAPIUrl, chain } = useNetworkConfigStore();
+  const { abi } = useABI(target);
   const { t } = useTranslation('common');
-  const { data: ensAddress } = useNetworkEnsAddress({
-    name: target?.toLowerCase(),
-    chainId: chain.id,
-  });
-  const client = useNetworkPublicClient();
-
-  useEffect(() => {
-    const loadABI = async () => {
-      if (target && ((ensAddress && isAddress(ensAddress)) || isAddress(target))) {
-        try {
-          const requestFunc = ({ method, params }: { method: any; params: any }) =>
-            client.request({ method, params });
-
-          const implementationAddress = await detectProxyTarget(ensAddress || target, requestFunc);
-
-          const response = await axios.get(
-            `${etherscanAPIUrl}&module=contract&action=getabi&address=${implementationAddress || ensAddress || target}`,
-          );
-          const responseData = response.data;
-
-          if (responseData.status === '1') {
-            const fetchedABI = JSON.parse(responseData.result);
-            setABI(fetchedABI);
-          } else {
-            setABI([]);
-          }
-        } catch (e) {
-          setABI([]);
-          logError(e, 'Error fetching ABI for smart contract');
-        }
-      } else {
-        setABI([]);
-      }
-    };
-    loadABI();
-  }, [target, ensAddress, etherscanAPIUrl, client]);
 
   /*
    * This makes component quite scoped to proposal / proposal template creation
